+++ src/nips/nip2.rs
@@ -1,4 +1,6 @@
-use crate::{events::EventPrepare, nostr_client::Client, utils::get_timestamp, Identity};
+use crate::{
+    events::EventPrepare, nostr_client::Client, req::ReqFilter, utils::get_timestamp, Identity,
+};
 
 // Implementation of the NIP2 protocol
 // https://github.com/nostr-protocol/nips/blob/master/02.md
@@ -69,5 +71,53 @@ impl Client {
 
         self.publish_event(&event)?;
         Ok(())
+    }
+
+    /// Get the contact list of a pub key
+    ///
+    /// # Example
+    /// ```rust
+    /// use nostr_rust::{nostr_client::Client, Identity, nips::nip2::ContactListTag};
+    /// use std::str::FromStr;
+    /// let mut client = Client::new(vec!["wss://nostr-pub.wellorder.net"]).unwrap();
+    /// let contact_list = client.get_contact_list("884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6").unwrap();
+    /// ```
+    pub fn get_contact_list(&mut self, pubkey: &str) -> Result<Vec<ContactListTag>, String> {
+        let mut contact_list: Vec<ContactListTag> = vec![];
+
+        let events = self.get_events_of(vec![ReqFilter {
+            ids: None,
+            authors: Some(vec![pubkey.to_string()]),
+            kinds: Some(vec![3]),
+            e: None,
+            p: None,
+            since: None,
+            until: None,
+            limit: Some(1),
+        }])?;
+
+        for event in events {
+            for tag in event.tags {
+                if tag[0] == "p" {
+                    let mut contact = ContactListTag {
+                        key: tag[1].clone(),
+                        main_relay: None,
+                        surname: None,
+                    };
+
+                    if tag.len() > 2 {
+                        contact.main_relay = Some(tag[2].clone());
+
+                        if tag.len() > 3 {
+                            contact.surname = Some(tag[3].clone());
+                        }
+                    }
+
+                    contact_list.push(contact);
+                }
+            }
+        }
+
+        Ok(contact_list)
     }
 }