+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.0 - NIP-02 Support
+
+- Added: `Client.get_contact_list` method
+- Added: `Client.add_event` method
+- Added: `Client.get_events` method
+- Added: `Client.get_events_of` method
+
 ## 0.2.0 - Architecture change
 
 - Removed: `Client.listen` function (Replaced by `Client.next_data`)